Update question1
Assignment-
Given a sorted array of "nums", remove the duplicates so each unique element appears only once.

You must have the result placed in the first part of the array nums. If there are k unique elements, then the first k elements of nums should hold the final result. The numbers beyond that should be unset/undefined/null keeping the length of the array to be the same.

Return k after placing the final result in the first k slots of nums.

Do not allocate extra space for another array.

For eg:

Example1:
Input: nums = [1,1,2]
Output: 2, nums = [1,2,_] - There are 2 unique elements so return 2 and the input array should be modified such that the unique elements are placed first.

Example2:
Input: nums = [0,0,1,1,1,2,2,3,3,4]
Output: 5, nums = [0,1,2,3,4,_,_,_,_,_] - There are 5 unique elements so return 5 and the input array should be modified such that the unique elements are placed first.